Move ConnectivityManager to :core:android:network module
