+++ core/android/network/src/main/kotlin/net/thunderbird/core/android/network/ConnectivityManagerApi23.kt
@@ -1,34 +1,47 @@
-package com.fsck.k9.network
+package net.thunderbird.core.android.network
 
 import android.net.ConnectivityManager.NetworkCallback
 import android.net.Network
 import android.net.NetworkCapabilities
+import android.net.NetworkRequest
 import android.os.Build
 import androidx.annotation.RequiresApi
 import timber.log.Timber
 import android.net.ConnectivityManager as SystemConnectivityManager
 
-@RequiresApi(Build.VERSION_CODES.N)
-internal class ConnectivityManagerApi24(
+@RequiresApi(Build.VERSION_CODES.M)
+internal class ConnectivityManagerApi23(
     private val systemConnectivityManager: SystemConnectivityManager,
 ) : ConnectivityManagerBase() {
     private var isRunning = false
-    private var isNetworkAvailable: Boolean? = null
+    private var lastActiveNetwork: Network? = null
+    private var wasConnected: Boolean? = null
 
     private val networkCallback = object : NetworkCallback() {
         override fun onAvailable(network: Network) {
             Timber.v("Network available: $network")
-            synchronized(this@ConnectivityManagerApi24) {
-                isNetworkAvailable = true
-                notifyOnConnectivityChanged()
-            }
+            notifyIfActiveNetworkOrConnectivityHasChanged()
         }
 
         override fun onLost(network: Network) {
             Timber.v("Network lost: $network")
-            synchronized(this@ConnectivityManagerApi24) {
-                isNetworkAvailable = false
-                notifyOnConnectivityLost()
+            notifyIfActiveNetworkOrConnectivityHasChanged()
+        }
+
+        private fun notifyIfActiveNetworkOrConnectivityHasChanged() {
+            val activeNetwork = systemConnectivityManager.activeNetwork
+            val isConnected = isNetworkAvailable()
+
+            synchronized(this@ConnectivityManagerApi23) {
+                if (activeNetwork != lastActiveNetwork || isConnected != wasConnected) {
+                    lastActiveNetwork = activeNetwork
+                    wasConnected = isConnected
+                    if (isConnected) {
+                        notifyOnConnectivityChanged()
+                    } else {
+                        notifyOnConnectivityLost()
+                    }
+                }
             }
         }
     }
@@ -38,7 +51,8 @@ internal class ConnectivityManagerApi24(
         if (!isRunning) {
             isRunning = true
 
-            systemConnectivityManager.registerDefaultNetworkCallback(networkCallback)
+            val networkRequest = NetworkRequest.Builder().build()
+            systemConnectivityManager.registerNetworkCallback(networkRequest, networkCallback)
         }
     }
 
@@ -52,12 +66,6 @@ internal class ConnectivityManagerApi24(
     }
 
     override fun isNetworkAvailable(): Boolean {
-        return synchronized(this) { isNetworkAvailable } ?: isNetworkAvailableSynchronous()
-    }
-
-    // Sometimes this will return 'true' even though networkCallback has already received onLost().
-    // That's why isNetworkAvailable() prefers the state derived from the callbacks over this method.
-    private fun isNetworkAvailableSynchronous(): Boolean {
         val activeNetwork = systemConnectivityManager.activeNetwork ?: return false
         val networkCapabilities = systemConnectivityManager.getNetworkCapabilities(activeNetwork)
         return networkCapabilities?.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET) == true
