+++ core/android/network/src/main/kotlin/net/thunderbird/core/android/network/KoinModule.kt
@@ -1,10 +1,10 @@
-package com.fsck.k9.network
+package net.thunderbird.core.android.network
 
 import android.content.Context
 import org.koin.dsl.module
 import android.net.ConnectivityManager as SystemConnectivityManager
 
-internal val connectivityModule = module {
+val coreAndroidNetworkModule = module {
     single { get<Context>().getSystemService(Context.CONNECTIVITY_SERVICE) as SystemConnectivityManager }
     single { ConnectivityManager(systemConnectivityManager = get()) }
 }
