+++ legacy/core/src/main/java/com/fsck/k9/CoreKoinModules.kt
@@ -11,13 +11,14 @@ import com.fsck.k9.mailstore.mailStoreModule
 import com.fsck.k9.message.extractors.extractorModule
 import com.fsck.k9.message.html.htmlModule
 import com.fsck.k9.message.quote.quoteModule
-import com.fsck.k9.network.connectivityModule
 import com.fsck.k9.notification.coreNotificationModule
 import com.fsck.k9.power.powerModule
 import com.fsck.k9.preferences.preferencesModule
+import net.thunderbird.core.android.network.coreAndroidNetworkModule
 
 val coreModules = listOf(
     mainModule,
+    coreAndroidNetworkModule,
     openPgpModule,
     autocryptModule,
     mailStoreModule,
@@ -30,7 +31,6 @@ val coreModules = listOf(
     jobModule,
     helperModule,
     preferencesModule,
-    connectivityModule,
     powerModule,
     loggingModule,
 )
